MacOS support
+ Added back optional features
+ misc changes

+++ blitz-app-adblock.js
@@ -9,9 +9,27 @@ var noUpdate = false;
 var autoGuest = false;
 
 const appPath = path.join(process.env.APPDATA, '..\\Local\\Programs\\Blitz\\resources');
- 
+
 async function start() {
     try {
+        var args = process.argv.slice(2);
+        for (var i = 0; i < args.length; i++) {
+            switch (args[i]) {
+            case '-noupdate':
+                noUpdate = true;
+                break;
+            case '-autoguest':
+                autoGuest = true;
+                break;
+            default:
+                console.log(`Unknown argument: '` + args[i] + `'`);
+            }
+        }
+        
+        if (process.platform === "darwin") {
+            appPath = '\\Applications\\Blitz.app\\Contents\\Resources';
+        }
+
         if(!fs.existsSync(`${appPath}\\app.asar`)) {
             Console.WriteLine("app.asar not found!");
         }
@@ -29,19 +47,24 @@ async function start() {
             console.log('Patching...');
     
             // copy adblocker lib to src
-            io.copyFile('adblocker.umd.min.js', `${appPath}\\app\\src\\adblocker.umd.min.js`)
+            if(fs.existsSync(`adblocker.umd.min.js`)) 
+                io.copyFile('adblocker.umd.min.js', `${appPath}\\app\\src\\adblocker.umd.min.js`)
+            else io.copyFile('./build/adblocker.umd.min.js', `${appPath}\\app\\src\\adblocker.umd.min.js`)
     
             // start writing our payload to createWindow.js
             io.modifyFileAtLine(js.filterEngine, `${appPath}\\app\\src\\createWindow.js`, 119);
             io.modifyFileAtLine('session: true,', `${appPath}\\app\\src\\createWindow.js`, 106);
     
             // optional features
-            if (noUpdate)  IO.ModifyFileAtLine('if (false) {', `${appPath}\\app\\src\\index.js`, 267);
-            if (autoGuest) IO.ModifyFileAtLine(js.autoGuest, `${appPath}\\app\\src\\preload.js`, 18);
+            if (noUpdate)  io.ModifyFileAtLine('if (false) {', `${appPath}\\app\\src\\index.js`, 267);
+            if (autoGuest) io.ModifyFileAtLine(js.autoGuest, `${appPath}\\app\\src\\preload.js`, 18);
     
-            // repack & cleanup
+            // repack
             console.log('Repacking app.asar...');
             await asar.createPackage(`${appPath}\\app\\`, `${appPath}\\app.asar`);
+
+            // cleanup
+            console.log('Cleaning up directory...');
             io.deleteFolder(`${appPath}\\app\\`);
             
             console.log('\r\nPatching complete! GLHF :)');
